refactor Cards to Card Component and add useToggle
