+++ package.json
@@ -10,6 +10,7 @@
     "firebase": "7.10.0",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
+    "react-hooks-lib": "0.1.5",
     "react-modal": "3.11.2",
     "react-native-gesture-handler": "1.6.0",
     "react-native-modal": "11.5.4",
@@ -17,7 +18,6 @@
     "react-native-swipeable-modal": "2.0.1",
     "react-scripts": "3.4.0",
     "react-swipeable": "5.5.1",
-    "react-uid": "2.2.0",
     "styled-components": "5.0.1",
     "uid": "1.0.0"
   },
